pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vulcan"
version = "0.1.0"
description = "Vulcan: a framework and backend services for vulnerability detection and dataset management."
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = { file = "LICENSE" }
authors = [
  { name = "NISL Vulcan Team" }
]
dependencies = [
  # Backend and API dependencies (from requirements_backend.txt and vulcan.yaml)
  "Flask==2.3.2",
  "Flask-CORS==4.0.0",
  "PyYAML==6.0.1",
  "requests==2.31.0",
  "psutil==5.9.8",

  # Training and graph-learning dependencies (from vulcan.yaml pip section)
  "absl-py==1.4.0",
  "aiohttp==3.9.0",
  "antlr4-python3-runtime>=4.9.3,<4.10",
  "antlr4-tools==0.2.1",
  "async-timeout==4.0.3",
  "attrs==23.1.0",
  "blinker==1.6.2",
  "bottleneck==1.3.5",
  "brotli==1.0.9",
  "certifi==2023.11.17",
  "cffi==1.16.0",
  "click==8.1.7",
  "cryptography==41.0.7",
  "daal4py>=2023.1.1",
  "datasets==2.12.0",
  # DGL 1.1.3 must be installed from official wheels before installing this package:
  #   pip install dgl -f https://data.dgl.ai/wheels/repo.html
  "dill==0.3.6",
  "filelock==3.13.1",
  "frozenlist==1.4.0",
  "fsspec==2023.10.0",
  "fst-pso==1.8.1",
  "fuzzytm==2.0.5",
  "gensim==4.3.0",
  "google-auth==2.22.0",
  "google-auth-oauthlib==0.5.2",
  "googledrivedownloader==0.4",
  "grpcio==1.48.2",
  "huggingface-hub==0.17.3",
  "idna==3.4",
  "importlib-metadata==7.0.1",
  "install-jdk==1.1.0",
  "isodate==0.6.1",
  "jinja2==3.1.2",
  "joblib==1.2.0",
  "markdown==3.4.1",
  "markupsafe==2.1.3",
  "miniful==0.0.6",
  "mkl-fft==1.3.8",
  "mkl-random==1.2.4",
  "mkl-service==2.4.0",
  "multidict==6.0.4",
  "multiprocess==0.70.14",
  "networkx==3.1",
  "numexpr==2.8.7",
  "numpy>=1.26.3,<1.27",
  "oauthlib==3.2.2",
  "omegaconf==2.3.0",
  "packaging==23.1",
  "pandas==2.1.4",
  "pillow==10.0.1",
  "programl==0.3.2",
  "protobuf==3.20.3",
  "pyarrow==11.0.0",
  "pyfume==0.2.25",
  "pyjwt==2.4.0",
  "pyopenssl==23.2.0",
  "pyparsing==3.0.9",
  "pysocks==1.7.1",
  "python-louvain==0.16",
  "pytz==2023.3.post1",
  "pyyaml==6.0.1",
  "rdflib==7.0.0",
  "regex==2023.10.3",
  "safetensors==0.4.0",
  "scikit-learn==1.3.0",
  "scikit-learn-intelex==2023.2.0",
  "scipy==1.11.4",
  "setuptools==68.2.2",
  "simpful==2.11.1",
  "smart-open==5.2.1",
  "tabulate==0.9.0",
  "tensorboard==2.12.1",
  "tensorboard-data-server==0.7.0",
  "tensorboard-plugin-wit==1.8.1",
  "tokenizers==0.13.3",
  "torch==1.11.0",
  "torch-cluster==1.6.0",
  "torch-geometric==1.7.2",
  "torch-scatter==2.0.9",
  "torch-sparse==0.6.15",
  "torchaudio==0.11.0",
  "torchvision==0.12.0",
  "tqdm==4.65.0",
  "transformers==4.32.1",
  "tree-sitter==0.20.4",
  "typing-extensions==4.9.0",
  "urllib3==1.26.18",
  "werkzeug>=2.3.3",
  "wheel==0.41.2",
  "xxhash==2.0.2",
  "yacs==0.1.8",
  "yarl==1.9.3",
  "zipp==3.17.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[project.scripts]
vulcan-train = "vulcan.cli.train:cli_main"
vulcan-val = "vulcan.cli.val:cli_main"
vulcan-benchmark = "vulcan.cli.benchmark:cli_main"
vulcan-export = "vulcan.cli.export:cli_main"
vulcan-backend = "vulcan.services.backend_server:run_backend"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
